Create README.md
In this project i decided to do a normal Simon game. It is a memory game where you have to follow the sequence and copy the sequence correctly to move on to the next level. I decided to have 2 modes in this game, medium and hard. The medium mode is pretty straight forward. When a coloured button flashes, you click on the button. When you get futher on in the game you will have several colours to rememeber in order. The game gets harder the longer you play! As for the hard mode, i decided to make the buttons change positions every time you move on to the next level, i thought this would make it harder for the user to rememeber each colour in order. For the design of the game i decided i wanted to have a big circle as the game board and 6 coloured buttons all around inside the big for the sequence. In this middle of the game board i have put a smaller circle with the control buttons in. This way the user can always see what level they are, on as well as being able to change the mode, start and restart the game. I decided to use bright and bold colours for the buttons as i thought they would stand out better from the dark grey game board. I also have created a modal with instuctions to my game. I thought it was best to add these so if someone was unsure on how to play the game, they can read the instructions and understand it. I would like to think the overall design of the game fits well with the common memory game.